Float Support

+++ src/main.rs
@@ -2,15 +2,15 @@ use std::io;
 
 fn main() {
     let mut input_str = String::new();
-    let mut input_num:usize =0;
+    let mut input_num:f64 =0.0;
     let mut conversion_index:usize = 1;
     let mut choice:&str;
     let mut types: Vec<&str>;
     let mut input_type: &str;
     let mut output_type: &str;
     let mut input_value = String::new();
-    let mut input_int: usize;
-    let mut output_value: usize = 0;
+    let mut input_int: f64;
+    let mut output_value: f64 =0.0;
 
     let conversions :[&str;6] = [
         "Celsius to Fahrenheit",
@@ -41,32 +41,32 @@ fn main() {
     input_value.pop();
     // println!("The length is {}",input_value.len());
     // println!("Output Type : {}",output_type);
-    input_int = input_value.parse::<usize>().unwrap();
+    input_int = input_value.parse::<f64>().unwrap();
 
     if input_type=="Celsius"{
         if output_type == "Fahrenheit"{
-            output_value = (input_int * 9/5)+32;
+            output_value = (input_int * 9.0/5.0)+32.0;
         }
         if output_type == "Kelvin"{
-            output_value = input_int + 273;
+            output_value = input_int + 273.0;
         }
     }else if input_type=="Fahrenheit"{
         if output_type == "Celsius"{
-            output_value = 5/9 *(input_int-32);
+            output_value = 5.0/9.0 *(input_int-32.0);
         }
         if output_type == "Kelvin"{
-            output_value = (5/9 *(input_int-32)) + 273;
+            output_value = (5.0/9.0 *(input_int-32.0)) + 273.0;
         }
     }else if input_type=="Kelvin"{
         if output_type == "Celsius"{
-            output_value = input_int - 273;
+            output_value = input_int - 273.0;
         }
         if output_type == "Fahrenheit"{
-            output_value = (    (input_int - 273)* 9/5)+32;
+            output_value = ((input_int - 273.0)* 9.0/5.0)+32.0;
         }
     }
     else {
-        output_value = 0;
+        output_value = 0.0;
     }
